Split out test_crawler_runners.py.
